# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: NestPro PG & Hostel OS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /properties/{propertyId}/buildings:
    get:
      operationId: listBuildings
      tags: [buildings]
      summary: List buildings in a property
      parameters:
        - name: propertyId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Building"
    post:
      operationId: createBuilding
      tags: [buildings]
      summary: Create a building
      parameters:
        - name: propertyId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BuildingInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Building"
  /buildings/{id}:
    patch:
      operationId: updateBuilding
      tags: [buildings]
      summary: Update a building
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BuildingUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Building"
    delete:
      operationId: deleteBuilding
      tags: [buildings]
      summary: Delete a building
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Floors ───────────────────────────────────────────────────
...
  /rooms/{roomId}/beds:
    get:
      operationId: listBeds
      tags: [beds]
      summary: List beds in a room
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bed"
    post:
      operationId: createBed
      tags: [beds]
      summary: Create a bed
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BedInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bed"
  /beds/{id}:
    patch:
      operationId: updateBed
      tags: [beds]
      summary: Update a bed
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BedUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bed"
    delete:
      operationId: deleteBed
      tags: [beds]
      summary: Delete a bed
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /properties/{propertyId}/beds:
    get:
      operationId: listPropertyBeds
      tags: [beds]
      summary: List all beds in a property (flat list with location info)
      parameters:
        - name: propertyId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BedWithLocation"

  # ─── Guests ───────────────────────────────────────────────────
...
  /complaints:
    get:
      operationId: listComplaints
      tags: [complaints]
      summary: List complaints
      parameters:
        - name: propertyId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, assigned, in_progress, resolved, closed]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Complaint"
    post:
      operationId: createComplaint
      tags: [complaints]
      summary: Create a complaint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComplaintInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Complaint"
  /complaints/{id}:
    get:
      operationId: getComplaint
      tags: [complaints]
      summary: Get a complaint
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Complaint"
    patch:
      operationId: updateComplaint
      tags: [complaints]
      summary: Update complaint status/assignment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComplaintUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Complaint"

  # ─── Staff ────────────────────────────────────────────────────
...
components:
  schemas:
...
    Building:
      type: object
      required: [id, propertyId, name, createdAt]
      properties:
        id:
          type: integer
        propertyId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    BuildingInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
    BuildingUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string

    # ─── Floor ───────────────────────────────────────────────────
...
    Bed:
      type: object
      required: [id, roomId, label, status, createdAt]
      properties:
        id:
          type: integer
        roomId:
          type: integer
        label:
          type: string
        status:
          type: string
          enum: [available, occupied, maintenance, reserved]
        monthlyRent:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time
    BedInput:
      type: object
      required: [label]
      properties:
        label:
          type: string
        status:
          type: string
          enum: [available, occupied, maintenance, reserved]
        monthlyRent:
          type: number
    BedUpdate:
      type: object
      properties:
        label:
          type: string
        status:
          type: string
          enum: [available, occupied, maintenance, reserved]
        monthlyRent:
          type: number
    BedWithLocation:
      type: object
      required: [id, roomId, label, status, roomNumber, floorName, buildingName]
      properties:
        id:
          type: integer
        roomId:
          type: integer
        label:
          type: string
        status:
          type: string
          enum: [available, occupied, maintenance, reserved]
        monthlyRent:
          type: ["number", "null"]
        roomNumber:
          type: string
        floorName:
          type: string
        buildingName:
          type: string
        guestName:
          type: ["string", "null"]
        guestId:
          type: ["integer", "null"]

    # ─── Guest ───────────────────────────────────────────────────
...
    Complaint:
      type: object
      required: [id, guestId, propertyId, title, category, status, priority, createdAt]
      properties:
        id:
          type: integer
        guestId:
          type: integer
        propertyId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
          enum: [maintenance, cleanliness, security, noise, wifi, water, electricity, food, other]
        status:
          type: string
          enum: [pending, assigned, in_progress, resolved, closed]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        assignedTo:
          type: ["integer", "null"]
        resolvedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
    ComplaintInput:
      type: object
      required: [guestId, propertyId, title, category, priority]
      properties:
        guestId:
          type: integer
        propertyId:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [maintenance, cleanliness, security, noise, wifi, water, electricity, food, other]
        priority:
          type: string
          enum: [low, medium, high, urgent]
    ComplaintUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, assigned, in_progress, resolved, closed]
        assignedTo:
          type: integer
        priority:
          type: string
          enum: [low, medium, high, urgent]
        description:
          type: string

    # ─── Staff ───────────────────────────────────────────────────
...