# lib/api-zod/src/generated/types/bedUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NestPro PG & Hostel OS API
 * OpenAPI spec version: 0.1.0
 */
import type { BedUpdateStatus } from './bedUpdateStatus';

export interface BedUpdate {
  label?: string;
  status?: BedUpdateStatus;
  monthlyRent?: number;
}